+++ packages/react-moveable/src/react-moveable/MoveableManager.tsx
@@ -184,7 +184,7 @@ export default class MoveableManager<T = {}>
      *     }
      * });
      */
-    public isMoveableElement(target: HTMLElement | SVGElement) {
+    public isMoveableElement(target: Element) {
         return target && ((target.getAttribute("class") || "").indexOf(PREFIX) > -1);
     }
     /**
@@ -610,7 +610,7 @@ export default class MoveableManager<T = {}>
     ) {
         const props = this.props as any;
         const triggerAblesSimultaneously = props.triggerAblesSimultaneously;
-        const enabledAbles = ables!.filter(able => able && props[able.name]);
+        const enabledAbles = ables!.filter(able => able && (able.always || props[able.name]));
 
         const dragStart = `drag${eventAffix}Start` as "dragStart";
         const pinchStart = `pinch${eventAffix}Start` as "pinchStart";